+++ vocprez/model/__init__.py
@@ -1,4 +1,5 @@
-from .catalog import CatalogRenderer
+from .vocprez import VocPrezRenderer
+from .vocabularies import VocabulariesRenderer
 from .vocabulary import VocabularyRenderer
 from .concept import ConceptRenderer
 from .collection import CollectionRenderer