+++ vocprez/model/profiles.py
@@ -13,6 +13,17 @@ profile_skos = Profile(
     default_language="en",
 )
 
+profile_vocpub = Profile(
+    "https://w3id.org/profile/vocpub",
+    label="VocPub",
+    comment="A profile of SKOS for the publication of Vocabularies. This profile mandates the use of one Concept "
+            "Scheme per vocabulary",
+    mediatypes=["text/html", "application/json"] + Renderer.RDF_MEDIA_TYPES,
+    default_mediatype="text/html",
+    languages=["en"],  # default 'en' only for now
+    default_language="en",
+)
+
 profile_dcat = Profile(
     "https://www.w3.org/TR/vocab-dcat/",
     label="DCAT",
@@ -56,5 +67,16 @@ profile_dd = Profile(
     mediatypes=["application/json"],
     default_mediatype="application/json",
     languages=["en"],
+    default_language="en",
+)
+
+profile_sdo = Profile(
+    "https://schema.org",
+    label="schema.org",
+    comment="Schema.org is a collaborative, community activity with a mission to create, maintain, and promote schemas "
+            "for structured data on the Internet, on web pages, in email messages, and beyond.",
+    mediatypes=["text/html"] + Renderer.RDF_MEDIA_TYPES,
+    default_mediatype="text/html",
+    languages=["en"],  # default 'en' only for now
     default_language="en",
 )